% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGE_GRP.R
\name{AGE_GRP}
\alias{AGE_GRP}
\title{Method of AGE_GRP function}
\usage{
AGE_GRP(dat)
}
\arguments{
\item{dat}{AGE(나이) 변수가 존재하는 data파일}
}
\value{
데이터에 AGE_GRP 군단연령 컬럼이 추가됨
}
\description{
AGE_GRP 함수는 Rawdata의 AGE 값을 이용하여 군단연령 컬럼을 추가하는 함수임
}
