% Generated by roxygen2 (4.0.1): do not edit by hand
\name{api_request}
\alias{api_request}
\title{Make a Goolge Analytics API request}
\usage{
api_request(url, token, simplify = TRUE, messages = FALSE)
}
\arguments{
\item{url}{the url of the page to retrieve.}

\item{token}{\code{Token2.0} class object.}

\item{simplify}{logical; should the result be simplified to a vector, matrix or data.frame.}

\item{messages}{logical; print information messages.}
}
\description{
Make a Goolge Analytics API request
}
\examples{
\dontrun{
    url <- "https://www.googleapis.com/analytics/v3/metadata/ga/olumns"
    data.json <- api_request(url)
}
}
\keyword{internal}

