% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dims_mets}
\alias{dims_mets}
\title{The shiny app to lists all columns for a Google Analytics given report type}
\usage{
dims_mets(data = ga)
}
\arguments{
\item{data}{dataset to show (now available only \code{ga}).}
}
\value{
\code{shiny.appobj} class object to show \code{\link{ga}} dataset.
}
\description{
The shiny app to lists all columns for a Google Analytics given report type
}

