% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny.R
\name{dimsmets}
\alias{dimsmets}
\title{The shiny app to lists all columns for a Google Analytics given report type}
\usage{
dimsmets(data = ga)
}
\arguments{
\item{data}{dataset to show (now available only \code{ga}).}
}
\value{
\code{shiny.appobj} class object to show \code{\link{ga}} dataset.
}
\description{
The shiny app to lists all columns for a Google Analytics given report type
}
\references{
\href{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}{Dimensions & Metrics Reference}
}
\seealso{
\code{\link{ga}} \code{\link{get_ga}}
}

