% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{get_custom_metric}
\alias{get_custom_metric}
\title{Custom Metrics}
\usage{
get_custom_metric(accountId, webPropertyId, customMetricId, token)
}
\arguments{
\item{accountId}{character. Account ID for the custom metric to retrieve.}

\item{webPropertyId}{character. Web property ID for the custom metric to retrieve.}

\item{customMetricId}{character. The ID of the custom metric to retrieve.}

\item{token}{\code{\link[httr]{Token2.0}} class object with a valid authorization data.}

\item{Path}{parameters Path parameters. Path parameters}
}
\value{
\item{kind}{Kind value for a custom metric. Set to "analytics#customMetric". It is a read-only field.}
\item{id}{Custom metric ID.}
\item{accountId}{Account ID.}
\item{webPropertyId}{Property ID.}
\item{name}{Name of the custom metric.}
\item{index}{Index of the custom metric.}
\item{scope}{Scope of the custom metric: HIT or PRODUCT.}
\item{active}{Boolean indicating whether the custom metric is active.}
\item{type}{Data type of custom metric.}
\item{min_value}{Min value of custom metric.}
\item{max_value}{Max value of custom metric.}
\item{created}{Time the custom metric was created.}
\item{updated}{Time the custom metric was last modified.}
}
\description{
Get a custom metric to which the user has access.
}
\references{
\href{https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/customMetrics}{Management API - Custom Metrics Overview}
}
\seealso{
Other Management API: \code{\link{get_custom_dimension}},
  \code{\link{get_experiment}}, \code{\link{get_filter}},
  \code{\link{get_goal}}, \code{\link{get_profile}},
  \code{\link{get_remarketing_audience}},
  \code{\link{get_unsampled_report}},
  \code{\link{get_upload}}, \code{\link{get_webproperty}},
  \code{\link{list_accounts}},
  \code{\link{list_custom_data_sources}},
  \code{\link{list_custom_dimensions}},
  \code{\link{list_custom_metrics}},
  \code{\link{list_experiments}},
  \code{\link{list_filters}}, \code{\link{list_goals}},
  \code{\link{list_profiles}},
  \code{\link{list_remarketing_audiences}},
  \code{\link{list_segments}},
  \code{\link{list_unsampled_reports}},
  \code{\link{list_uploads}},
  \code{\link{list_webproperties}}
}

