% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_firstdate}
\alias{get_firstdate}
\title{Get the first date with available data}
\usage{
get_firstdate(profile.id, token)
}
\arguments{
\item{profile.id}{Google Analytics profile ID.}

\item{token}{\code{Token2.0} class object.}
}
\value{
date in YYYY-MM-DD format.
}
\description{
Get the first date with available data
}
\examples{
\dontrun{
first.date <- get_firstdate(profile.id = "myProfileID", token = ga_token)
}
}

