% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_metadata}
\alias{get_metadata}
\title{Lists all columns for a report type}
\usage{
get_metadata(report.type = "ga")
}
\arguments{
\item{report.type}{report type (now allowed only 'ga').}
}
\value{
A data frame with metadata information.
}
\description{
Lists all columns for a report type
}
\examples{
ga_meta <- get_metadata()
names(ga_meta)
table(ga_meta$status)
table(ga_meta$type)
subset(ga_meta, subset = group == "User" & type == "METRIC" & status != "DEPRECATED", select = ids)
}
\references{
Dimensions & Metrics Reference: \url{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}
}

