% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_profiles}
\alias{get_profiles}
\title{Lists views (profiles) to which the user has access}
\usage{
get_profiles(token, account.id = "~all", webproperty.id = "~all",
  start.index = 1L, max.results = 1000L)
}
\arguments{
\item{token}{\code{Token2.0} class object.}

\item{account.id}{account ID for the view (profiles) to retrieve. Can either be a specific account ID or "~all", which refers to all the accounts to which the user has access.}

\item{webproperty.id}{web property ID for the views (profiles) to retrieve. Can either be a specific web property ID or "~all", which refers to all the web properties to which the user has access.}

\item{start.index}{an index of the first account to retrieve.}

\item{max.results}{the maximum number of accounts to include in this response.}
}
\value{
A data frame with Google Analytics management data.
}
\description{
Lists views (profiles) to which the user has access
}
\seealso{
\code{\link{get_token}}

Other Management.API: \code{\link{get_accounts}};
  \code{\link{get_goals}}; \code{\link{get_segments}};
  \code{\link{get_webproperties}}
}

