% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_report}
\alias{get_report}
\title{Get Google Anaytics report data}
\usage{
get_report(query, token, date.format = "\%Y-\%m-\%d", messages = FALSE)
}
\arguments{
\item{query}{\code{GAQuery} class object.}

\item{token}{\code{Token2.0} class object.}

\item{date.format}{date format.}

\item{messages}{print information messages.}
}
\value{
A data frame with Google Analytics reporting data.
}
\description{
Get Google Anaytics report data
}

