% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_unsampled}
\alias{get_unsampled}
\title{Lists unsampled reports which the user has access to}
\usage{
get_unsampled(account.id, webproperty.id, profile.id, start.index = NULL,
  max.results = NULL, token, verbose = getOption("rga.verbose", FALSE))
}
\arguments{
\item{account.id}{integer or character. The account ID to retrieve unsampled reports for. Must be a specific account ID, ~all is not supported.}

\item{webproperty.id}{character. Web property ID to retrieve unsampled reports for. Must be a specific web property ID, ~all is not supported. Requires specified \code{account.id}.}

\item{profile.id}{ineger or character. View (Profile) ID to retrieve unsampled reports for. Must be a specific view (profile) ID, ~all is not supported. Requires specified \code{account.id} and \code{webproperty.id}.}

\item{start.index}{integer. An index of the first unsampled report to retrieve. Use this parameter as a pagination mechanism along with the max-results parameter.}

\item{max.results}{integer. The maximum number of unsampled reports to include in this response.}

\item{token}{\code{\link[httr]{Token2.0}} class object with a valid authorization data.}

\item{verbose}{logical. Should print information verbose?}
}
\value{
\item{id}{Unsampled report ID.}
\item{title}{Title of the unsampled report.}
\item{accountId}{Account ID to which this unsampled report belongs.}
\item{webPropertyId}{Web property ID to which this unsampled report belongs. The web property ID is of the form UA-XXXXX-YY.}
\item{profileId}{View (Profile) ID to which this unsampled report belongs.}
\item{start-date}{The start date for the unsampled report.}
\item{end-date}{The end date for the unsampled report.}
\item{metrics}{The metrics for the unsampled report.}
\item{dimensions}{The dimensions for the unsampled report.}
\item{filters}{The filters for the unsampled report.}
\item{segment}{The segment for the unsampled report.}
\item{status}{Status of this unsampled report. Possible values are PENDING, COMPLETED, or FAILED.}
\item{downloadType}{The type of download you need to use for the report data file.}
\item{driveDownloadDetails.documentId}{Id of the document/file containing the report data.}
\item{cloudStorageDownloadDetails.bucketId}{Id of the bucket the file object is stored in.}
\item{cloudStorageDownloadDetails.objectId}{Id of the file object containing the report data.}
\item{created}{Time this unsampled report was created.}
\item{updated}{Time this unsampled report was last modified.}
}
\description{
Lists unsampled reports which the user has access to
}
\references{
\href{https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/unsampledReports}{Google Management API - Unsampled Reports}
}
\seealso{
\code{\link{authorize}}

Other The Google Analytics Management API: \code{\link{get_accounts}};
  \code{\link{get_goals}}; \code{\link{get_profiles}};
  \code{\link{get_segments}};
  \code{\link{get_webproperties}}
}

