\name{rga-package}
\alias{rga-package}
\alias{rga}
\docType{package}
\title{
R Google Analytics
}
\description{
This is a package for extracting data from Google Analytics into R. The package uses OAuth 2.0 (protocol) to access the Google Analytics API.
}
\details{
\tabular{ll}{
Package: \tab rga\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2012-12-05\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Maintainer: Bror Skardhamar <bror@skardhamar.com>
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~
%~~ directory ~~
\keyword{
	rga
	google analytics
	google analytics api
}
\seealso{
	\code{\link[rga:rga.open]{rga.open}}
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
%\examples{
%  rga.open(instance="ga")
%}
