% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{RGA}
\alias{RGA}
\alias{RGA-package}
\alias{rga}
\title{A Google Analytics client for R}
\description{
This is a package for extracting data from Google Analytics into R. The package uses OAuth 2.0 (protocol) to access the Google Analytics API.
}
\references{
The Management API: \url{https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/}
The Core Reporting API: \url{https://developers.google.com/analytics/devguides/reporting/core/v3/reference}
The Multi-Channel Funnels Reporting API: \url{https://developers.google.com/analytics/devguides/reporting/mcf/v3/reference}
}
\keyword{package}

