% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shiny_demo}
\alias{shiny_demo}
\title{The shiny app to lists all columns for a Google Analytics given report type}
\usage{
shiny_demo(data = ga)
}
\arguments{
\item{data}{dataset to show (now available only \code{\link{ga}}).}
}
\description{
The shiny app to lists all columns for a Google Analytics given report type
}

