% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show_dimsmets}
\alias{show_dimsmets}
\title{The shiny app to lists all columns for a Google Analytics given report type}
\usage{
show_dimsmets(data = ga)
}
\arguments{
\item{data}{dataset to show (now available only \code{\link{ga}}).}
}
\value{
\code{shiny.appobj} class object to show \link{\code{ga}} dataset.
}
\description{
The shiny app to lists all columns for a Google Analytics given report type
}

