% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test_query}
\alias{test_query}
\title{Checking Google Analytics report query}
\usage{
test_query(type = c("ga", "mcf"), query, token, results = FALSE,
  messages = FALSE)
}
\arguments{
\item{type}{character string including report type. "ga" for core report, "mcf" for multi-channel funnels report.}

\item{query}{\code{GAQuery} class object including a request parameters.}

\item{token}{\code{\link[httr]{Token2.0}} class object with a valid authorization data.}

\item{results}{logical. Return the results or logical?}

\item{messages}{logical. Should print information messages?}
}
\description{
Checking Google Analytics report query
}
\examples{
\dontrun{
query <- set_query("ga:00000000")
test_qeury(query)
}
}

