% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny.R
\name{dimsmets}
\alias{dimsmets}
\title{The shiny app to lists all columns for a Google Analytics given report type}
\usage{
dimsmets(report.type = "ga")
}
\arguments{
\item{report.type}{character. Report type. Allowed Values: ga. Where ga corresponds to the Core Reporting API.}
}
\value{
\code{shiny.appobj} class object to show \code{\link{list_metadata}} output.
}
\description{
The shiny app to lists all columns for a Google Analytics given report type
}
\references{
\href{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}{Dimensions & Metrics Reference}
}
\seealso{
\code{\link{list_metadata}} \code{\link{get_ga}}
}

