% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_accounts}
\alias{get_accounts}
\title{Lists all accounts which the user has access to}
\usage{
get_accounts(start.index = NULL, max.results = NULL, token,
  verbose = getOption("rga.verbose", FALSE))
}
\arguments{
\item{start.index}{integer. An index of the first account to retrieve. Use this parameter as a pagi- nation mechanism along with the max-results parameter.}

\item{max.results}{integer. The maximum number of accounts to include in this response.}

\item{token}{\code{\link[httr]{Token2.0}} class object with a valid authorization data.}

\item{verbose}{logical. Should print information verbose?}
}
\value{
An account collection provides a list of Analytics accounts to which a user has access. The account collection is the entry point to all management information. Each resource in the collection corresponds to a single Analytics account.
\item{id}{Account ID.}
\item{name}{Account name.}
\item{permissions}{All the permissions that the user has for this account. These include any implied permissions (e.g., EDIT implies VIEW).}
\item{created}{Time the account was created.}
\item{updated}{Time the account was last modified.}
}
\description{
Lists all accounts which the user has access to
}
\references{
\href{https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/accounts}{Google Management API - Accounts}
}
\seealso{
\code{\link{authorize}}

Other The Google Analytics Management API: \code{\link{get_custom_sources}};
  \code{\link{get_goals}}; \code{\link{get_profiles}};
  \code{\link{get_segments}}; \code{\link{get_unsampled}};
  \code{\link{get_webproperties}}
}

