% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_accounts}
\alias{get_accounts}
\title{Lists all accounts to which the user has access}
\usage{
get_accounts(token, start.index = 1L, max.results = 1000L)
}
\arguments{
\item{token}{\code{Token2.0} class object.}

\item{start.index}{an index of the first account to retrieve.}

\item{max.results}{the maximum number of accounts to include in this response.}
}
\value{
A data frame with Google Analytics management data.
\item{id}{account ID.}
\item{name}{account name.}
\item{effective}{all the permissions that the user has for this account.}
\item{created}{time the account was created.}
\item{updated}{time the account was last modified.}
}
\description{
Lists all accounts to which the user has access
}
\seealso{
\code{\link{get_token}}

Other Management.API: \code{\link{get_goals}};
  \code{\link{get_profiles}}; \code{\link{get_segments}};
  \code{\link{get_webproperties}}
}

