% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report.R
\name{get_firstdate}
\alias{get_firstdate}
\title{Get the first date with available data}
\usage{
get_firstdate(profile.id, token, verbose = getOption("rga.verbose", FALSE))
}
\arguments{
\item{profile.id}{Google Analytics profile ID. Can be obtained using the \code{\link{list_profiles}} or via the web interface Google Analytics.}

\item{token}{\code{\link[httr]{Token2.0}} class object with a valid authorization data.}

\item{verbose}{logical. Should print information verbose?}
}
\value{
Start date of collecting the Google Analytics statistics.
}
\description{
Get the first date with available data
}
\examples{
\dontrun{
authorize(client.id = "myID", client.secret = "mySecret")
first.date <- get_firstdate(profile.id = "myProfileID")
}
}
\seealso{
\code{\link{authorize}}

Other The Google Analytics Reporting API: \code{\link{get_ga}};
  \code{\link{get_mcf}}; \code{\link{get_rt}}
}

