% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_firstdate}
\alias{get_firstdate}
\title{Get the first date with available data}
\usage{
get_firstdate(profile.id, token, verbose = getOption("rga.verbose", FALSE))
}
\arguments{
\item{profile.id}{Google Analytics profile ID.}

\item{token}{\code{\link[httr]{Token2.0}} class object with a valid authorization data.}

\item{verbose}{logical. Should print information verbose?}
}
\value{
Start date of collecting the Google Analytics statistics.
}
\description{
Get the first date with available data
}
\examples{
\dontrun{
authorize(client.id = "myID", client.secret = "mySecret")
first.date <- get_firstdate(profile.id = "myProfileID")
}
}
\seealso{
\code{\link{authorize}}

Other The.Google.Analytics.Reporting.API: \code{\link{get_ga}};
  \code{\link{get_mcf}}; \code{\link{get_report}};
  \code{\link{get_rt}}
}

