% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_goals}
\alias{get_goals}
\title{Lists goals to which the user has access}
\usage{
get_goals(token, account.id = "~all", webproperty.id = "~all",
  profile.id = "~all", start.index = NULL, max.results = NULL)
}
\arguments{
\item{token}{\code{\link[httr]{Token2.0}} class object.}

\item{account.id}{account ID to retrieve goals for. Can either be a specific account ID or "~all", which refers to all the accounts that user has access to.}

\item{webproperty.id}{web property ID to retrieve goals for. Can either be a specific web property ID or "~all", which refers to all the web properties that user has access to.}

\item{profile.id}{view (Profile) ID to retrieve goals for. Can either be a specific view (profile) ID or "~all", which refers to all the views (profiles) that user has access to.'}

\item{start.index}{an index of the first goals to retrieve.}

\item{max.results}{the maximum number of goal to include in this response.}
}
\value{
A data frame with Google Analytics management data.
\item{id}{goal ID.}
\item{accountId}{account ID to which this goal belongs.}
\item{webPropertyId}{web property ID to which this goal belongs. The web property ID is of the form UA-XXXXX-YY.}
\item{profileId}{view (Profile) ID to which this goal belongs.}
\item{name}{goal name.}
\item{value}{goal value.}
\item{active}{determines whether this goal is active.}
\item{type}{goal type. Acceptable values are: "EVENT", "URL_DESTINATION", "VISIT_NUM_PAGES", "VISIT_TIME_ON_SITE"}
\item{created}{time this goal was created.}
\item{updated}{time this goal was last modified.}
}
\description{
Lists goals to which the user has access
}
\references{
Google Management API - Goals: \url{https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/goals}
}
\seealso{
\code{\link{authorize}}

Other Management.API: \code{\link{get_accounts}};
  \code{\link{get_profiles}}; \code{\link{get_segments}};
  \code{\link{get_webproperties}}
}

