% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_profiles}
\alias{get_profiles}
\title{Lists views (profiles) to which the user has access}
\usage{
get_profiles(token, account.id = "~all", webproperty.id = "~all",
  start.index = NULL, max.results = NULL)
}
\arguments{
\item{token}{\code{Token2.0} class object.}

\item{account.id}{account ID for the view (profiles) to retrieve. Can either be a specific account ID or "~all", which refers to all the accounts to which the user has access.}

\item{webproperty.id}{web property ID for the views (profiles) to retrieve. Can either be a specific web property ID or "~all", which refers to all the web properties to which the user has access.}

\item{start.index}{an index of the first account to retrieve.}

\item{max.results}{the maximum number of accounts to include in this response.}
}
\value{
A data frame with Google Analytics management data.
\item{id}{view (Profile) ID.}
\item{accountId}{account ID to which this view (profile) belongs.}
\item{webPropertyId}{web property ID of the form UA-XXXXX-YY to which this view (profile) belongs.}
\item{name}{name of this view (profile).}
\item{currency}{the currency type associated with this view (profile).}
\item{timezone}{time zone for which this view (profile) has been configured.}
\item{websiteUrl}{website URL for this view (profile).}
\item{type}{view (Profile) type. Supported types: WEB or APP.}
\item{eCommerceTracking}{indicates whether ecommerce tracking is enabled for this view (profile).}
\item{siteSearchQueryParameters}{the site search query parameters for this view (profile).}
\item{stripSiteSearchQueryParameters}{site search category parameters for this view (profile).}
\item{created}{time this view (profile) was created.}
\item{updated}{time this view (profile) was last modified.}
}
\description{
Lists views (profiles) to which the user has access
}
\references{
Views (Profiles): \url{https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/profiles}
}
\seealso{
\code{\link{get_token}}

Other Management.API: \code{\link{get_accounts}};
  \code{\link{get_goals}}; \code{\link{get_segments}};
  \code{\link{get_webproperties}}
}

