% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_report}
\alias{get_report}
\title{Get Google Anaytics report data}
\usage{
get_report(profile.id, start.date = "7daysAgo", end.date = "yesterday",
  metrics = "ga:users,ga:sessions,ga:pageviews", dimensions = "ga:date",
  sort = NULL, filters = NULL, segment = NULL, start.index = 1L,
  max.results = 10000L, date.format = "\%Y-\%m-\%d", query, token,
  messages = FALSE)
}
\arguments{
\item{profile.id}{Google Analytics profile ID. Can be character (with or without "ga:" prefix) or integer.}

\item{start.date}{start date for fetching Analytics data in YYYY-MM-DD format.}

\item{end.date}{rnd date for fetching Analytics data in YYYY-MM-DD format.}

\item{metrics}{a comma-separated list of Analytics metrics, such as ga:sessions,ga:bounces.}

\item{dimensions}{a comma-separated list of Analytics dimensions, such as ga:browser,ga:city.}

\item{sort}{a comma-separated list of dimensions or metrics that determine the sort order for Analytics data.}

\item{filters}{a comma-separated list of dimension or metric filters to be applied to Analytics data.}

\item{segment}{an Analytics segment to be applied to data.}

\item{start.index}{an index of the first entity to retrieve.}

\item{max.results}{the maximum number of entries to include in this feed.}

\item{token}{\code{Token2.0} class object.}

\item{date.format}{date format for output data.}

\item{messages}{print information messages.}

\item{query}{\code{GAQuery} class object.}
}
\value{
A data frame with Google Analytics reporting data.
}
\description{
\code{get_report} provide a query the Core or Multi-Channel Funnels Reporting API for Google Analytics report data.
}
\references{
Core Reporting API - Dimensions & Metrics Reference: \url{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}

Multi-Channel Funnels Reporting API - Dimensions & Metrics Reference: \url{https://developers.google.com/analytics/devguides/reporting/mcf/dimsmets/}

Google Analytics Query Explorer 2: \url{https://ga-dev-tools.appspot.com/explorer/}
}
\seealso{
\code{\link{get_token}}
}

