% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_report}
\alias{get_report}
\alias{get_report.GAQuery}
\alias{get_report.default}
\title{Get Google Anaytics report data}
\usage{
get_report(...)

\method{get_report}{default}(profile.id, start.date = "7daysAgo",
  end.date = "yesterday", metrics = "ga:users,ga:sessions,ga:pageviews",
  dimensions = "ga:date", sort = NULL, filters = NULL, segment = NULL,
  start.index = 1L, max.results = 10000L, token,
  date.format = "\%Y-\%m-\%d", messages = FALSE)

\method{get_report}{GAQuery}(query, token, date.format = "\%Y-\%m-\%d",
  messages = FALSE)
}
\arguments{
\item{profile.id}{Google Analytics profile ID. Can be character (with or without "ga:" prefix) or integer.}

\item{start.date}{start date for fetching Analytics data in YYYY-MM-DD format.}

\item{end.date}{rnd date for fetching Analytics data in YYYY-MM-DD format.}

\item{metrics}{a comma-separated list of Analytics metrics, such as ga:sessions,ga:bounces.}

\item{dimensions}{a comma-separated list of Analytics dimensions, such as ga:browser,ga:city.}

\item{sort}{a comma-separated list of dimensions or metrics that determine the sort order for Analytics data.}

\item{filters}{a comma-separated list of dimension or metric filters to be applied to Analytics data.}

\item{segment}{an Analytics segment to be applied to data.}

\item{start.index}{an index of the first entity to retrieve.}

\item{max.results}{the maximum number of entries to include in this feed.}

\item{token}{\code{Token2.0} class object.}

\item{date.format}{date format.}

\item{messages}{print information messages.}

\item{query}{\code{GAQuery} class object.}
}
\value{
A data frame with Google Analytics reporting data.
}
\description{
Get Google Anaytics report data




}
\seealso{
\code{\link{set_query}} \code{\link{get_token}}
}

