% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_segments}
\alias{get_segments}
\title{Lists segments to which the user has access}
\usage{
get_segments(token, start.index = 1L, max.results = 1000L)
}
\arguments{
\item{token}{\code{Token2.0} class object.}

\item{start.index}{an index of the first account to retrieve.}

\item{max.results}{the maximum number of accounts to include in this response.}
}
\value{
A data frame with Google Analytics management data.
\item{id}{segment ID.}
\item{segmentId}{segment ID. Can be used with the segment parameter in Data Feed.}
\item{name}{segment name.}
\item{definition}{segment definition.}
\item{type}{type for a segment. Possible values are "BUILT_IN" or "CUSTOM".}
\item{created}{time the segment was created.}
\item{updated}{time the segment was last modified.}
}
\description{
Lists segments to which the user has access
}
\references{
Segments: \url{https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/segments}
}
\seealso{
\code{\link{get_token}}

Other Management.API: \code{\link{get_accounts}};
  \code{\link{get_goals}}; \code{\link{get_profiles}};
  \code{\link{get_webproperties}}
}

