% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_token}
\alias{get_token}
\title{Generate an oauth2.0 token}
\usage{
get_token(client.id, client.secret, ...)
}
\arguments{
\item{client.id}{OAuth client ID.}

\item{client.secret}{OAuth client secret.}

\item{...}{arguments to be passed to \code{oauth2.0_token}.}
}
\value{
A Token2.0 reference class (RC) object.
}
\description{
\code{get_token} is wrapper for \code{link[httr]{oauth2.0_token} function}.
}
\details{
Client ID and client secret also can be set with \code{RGA_CONSUMER_ID} and \code{RGA_CONSUMER_SECRET} environment variables.
}
\examples{
\dontrun{
ga_token <- get_token(client.id = "myID", client.secret = "mySecret")
# if set RGA_CONSUMER_ID and RGA_CONSUMER_SECRET environment variables
ga_token <- get_token()
}
}
\references{
Google Developers Console: \url{https://console.developers.google.com/}

Environment variable: \url{http://en.wikipedia.org/wiki/Environment_variable}
}
\seealso{
Other OAuth: \code{\link[httr]{oauth_app}} \code{\link[httr]{oauth2.0_token}} \code{\link[httr]{Token-class}}

Setup environment variables: \code{\link{Startup}}
}

