% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_webproperties}
\alias{get_webproperties}
\title{Gets a web property to which the user has access}
\usage{
get_webproperties(token, account.id = "~all", start.index = 1L,
  max.results = 1000L)
}
\arguments{
\item{token}{\code{Token2.0} class object.}

\item{account.id}{account ID to retrieve web properties for. Can either be a specific account ID or "~all", which refers to all the accounts that user has access to.}

\item{start.index}{an index of the first account to retrieve.}

\item{max.results}{the maximum number of accounts to include in this response.}
}
\value{
A data frame with Google Analytics management data.
}
\description{
Gets a web property to which the user has access
}

