% Generated by roxygen2 (4.0.1): do not edit by hand
\name{help_ga}
\alias{help_ga}
\title{Getting help Google Analytics metadata}
\usage{
help_ga(topic, data = ga, type = c("all", "metric", "dimension"),
  depricated = FALSE, select = c("id", "uiName", "description"))
}
\arguments{
\item{topic}{a name or character string specifying the topic for which help is sought.}

\item{data}{a dataset for search (now available only 'ga').}

\item{type}{parameter type; metrics, dimensions or all.}

\item{depricated}{logical; print depricated parameters or not.}

\item{select}{columns to be returned. "all" return all columns}
}
\value{
Subset \code{\link{ga}} dataset accordingly search.
}
\description{
Getting help Google Analytics metadata
}
\examples{
help_ga("visit")
help_ga("session", type = "metric", select = "id")
}
\seealso{
\code{\link{ga}}
}

