% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_query}
\alias{set_query}
\title{Set Google Analytics report query}
\usage{
set_query(profile.id, start.date = "7daysAgo", end.date = "yesterday",
  metrics = "ga:users,ga:sessions,ga:pageviews", dimensions = "ga:date",
  sort = NULL, filters = NULL, segment = NULL, start.index = 1L,
  max.results = 10000L)
}
\arguments{
\item{profile.id}{Google Analytics profile ID. Can be character (with or without "ga:" prefix) or integer.}

\item{start.date}{start date for fetching Analytics data in YYYY-MM-DD format.}

\item{end.date}{rnd date for fetching Analytics data in YYYY-MM-DD format.}

\item{metrics}{a comma-separated list of Analytics metrics, such as ga:sessions,ga:bounces.}

\item{dimensions}{a comma-separated list of Analytics dimensions, such as ga:browser,ga:city.}

\item{sort}{a comma-separated list of dimensions or metrics that determine the sort order for Analytics data.}

\item{filters}{a comma-separated list of dimension or metric filters to be applied to Analytics data.}

\item{segment}{an Analytics segment to be applied to data.}

\item{start.index}{an index of the first entity to retrieve.}

\item{max.results}{the maximum number of entries to include in this feed.}
}
\value{
\code{GAQuery} class object.
}
\description{
\code{set_query} provide a query the Core or Multi-Channel Funnels Reporting API for Google Analytics report data.
}
\examples{
set_query(profile.id = "ga:00000000")
set_query(profile.id = "ga:00000000", start.date = "7daysAgo", end.date = "yesterday",
          metrics = "ga:users,ga:sessions,ga:pageviews", dimensions = "ga:date")
query <- set_query(profile.id = "ga:00000000", start.date = "31daysAgo", end.date = "yesterday",
                   metrics = "ga:sessions,ga:pageviews", dimensions = "ga:source,ga:medium")
query
query$sort <- "-ga:sessions"
query
query$sort <- NULL
query
}
\references{
Core Reporting API - Dimensions & Metrics Reference: \url{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}

Multi-Channel Funnels Reporting API - Dimensions & Metrics Reference: \url{https://developers.google.com/analytics/devguides/reporting/mcf/dimsmets/}

Google Analytics Query Explorer 2: \url{https://ga-dev-tools.appspot.com/explorer/}
}

