% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{shiny_dimsmets}
\alias{shiny_dimsmets}
\title{The Shiny Dimensions & Metrics Explorer}
\usage{
shiny_dimsmets()
}
\description{
The dimensions and metrics explorer lists and describes all the dimensions and metrics available through the Core Reporting API. This app deployed to the \url{https://artemklevtsov.shinyapps.io/ga-dimsmets}.
}
\seealso{
\code{\link{list_dimsmets}} \code{\link{get_ga}}
}

